Restructured directories and fixed token issues
