--- .prettierrc.json
@@ -1,5 +1,0 @@
-{
-	"tabWidth": 4,
-	"singleQuote": true,
-	"semi": true
-}
