+++ app/models.py
@@ -16,9 +16,9 @@ class User(UserMixin, db.Model):
     username: so.Mapped[str] = so.mapped_column(sa.String(64), index=True, unique=True)
     email: so.Mapped[str] = so.mapped_column(sa.String(120), index=True, unique=True)
     password_hash: so.Mapped[Optional[str]] = so.mapped_column(sa.String(256)) #remove nullable later
-    
+
     posts: so.WriteOnlyMapped['Post'] = so.relationship(back_populates='author')
-    
+
     def set_password(self, password):
         self.password_hash = generate_password_hash(password)
     def check_password(self, password):
@@ -35,10 +35,10 @@ class Post(db.Model):
     rating: so.Mapped[int] = so.mapped_column(sa.Integer())
     author: so.Mapped[User] = so.relationship(back_populates='posts')
     #comments: so.WriteOnlyMapped['Post'] = so.relationship(back_populates='post')
-    
+
     def __rpr__(self):
         return '<Post []>'.format(self.body)
-    
+
 # class Comment(db.Model):
 #     id: so.Mapped[int] = so.mapped_column(primary_key=True)
 #     user_id: so.Mapped[str] = so.mapped_column(sa.ForeignKey(User.id), index=False)
@@ -47,7 +47,7 @@ class Post(db.Model):
 #     post: so.Mapped[User] = so.relationship(back_populates='comments')
 #     comment: so.Mapped[str] = so.mapped_column(sa.String(140))
 #     timestamp: so.Mapped[datetime] = so.mapped_column(index=True, default=lambda: datetime.now(timezone.utc))
-    
+
 
 # @login.user_loader
 # def load_user(id):
