--- app/templates/testLogin.html
@@ -1,50 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <link href="//maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" rel="stylesheet" id="bootstrap-css">
-<script src="//maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
-<script src="//cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.3.1/css/all.css" integrity="sha384-mzrmE5qonljUremFsqc01SB46JvROS7bZs3IO2EmfFsd15uHvIt+Y8vEf7N7fWAU" crossorigin="anonymous">
-    <link rel="stylesheet" href="/style.css">
-    <title>Bootstrap 4 Login/Register Form</title>
-</head>
-
-  <body>
-      <div id="login">
-          <h3 class="text-center text-white pt-5">Login form</h3>
-          <div class="container">
-              <div id="login-row" class="row justify-content-center align-items-center">
-                  <div id="login-column" class="col-md-6">
-                      <div id="login-box" class="col-md-12">
-                          <form id="login-form" class="form" action="" method="post">
-                              <h3 class="text-center text-info">Hi, {{user.username}}</h3>
-                              <div class="form-group">
-                                  <label for="email" class="text-info">Email:</label><br>
-                                  <input type="text" name="email" id="email" class="form-control">
-                              </div>
-                              <div class="form-group">
-                                  <label for="password" class="text-info">Password:</label><br>
-                                  <input type="text" name="password" id="password" class="form-control">
-                              </div>
-                              <div class="form-group">
-                                  <label for="remember-me" class="text-info"><span>Remember me</span> <span><input id="remember-me" name="remember-me" type="checkbox"></span></label><br>
-                                  <input type="submit" name="submit" class="btn btn-info btn-md" value="submit">
-                                  {% for post in posts %}
-                                  <div><p>{{ post.author.username }} says: <b>{{ post.body }}</b></p></div>
-                                  {% endfor %}
-                              </div>
-                              <div id="register-link" class="text-right">
-                                  <a href="#" class="text-info">Register here</a>
-                              </div>
-                          </form>
-                      </div>
-                  </div>
-              </div>
-          </div>
-      </div>
-  </body>
-
