+++ requirements.txt
@@ -1,25 +1,68 @@
 alembic==1.13.1
+awscli==1.29.17
 bcrypt==4.1.2
 blinker==1.7.0
+boto3==1.28.17
+botocore==1.31.17
+certifi==2020.6.20
+chardet==3.0.4
 click==8.1.7
+colorama==0.4.4
+cycler==0.10.0
+docutils==0.16
+exceptiongroup==1.2.1
 Flask==3.0.3
 Flask-Bcrypt==1.0.1
 Flask-JWT-Extended==4.6.0
+Flask-Login==0.6.3
 Flask-Migrate==4.0.7
 Flask-SQLAlchemy==3.1.1
 Flask-WTF==1.2.1
 greenlet==3.0.3
+idna==2.10
+importlib_metadata==7.1.0
+importlib_resources==6.4.0
 iniconfig==2.0.0
+install==1.3.5
 itsdangerous==2.2.0
 Jinja2==3.1.3
+jmespath==1.0.1
+kiwisolver==1.2.0
 Mako==1.3.3
 MarkupSafe==2.1.5
+matplotlib==3.3.2
+mpmath==1.1.0
+msgpack==1.0.7
+Naked==0.1.32
+numpy==1.19.2
 packaging==24.0
+panda==0.3.1
+pandas==1.1.3
+Pillow==8.0.0
 pluggy==1.5.0
+pyasn1==0.5.0
+pycrypto==2.6.1
+pycryptodome==3.18.0
 PyJWT==2.8.0
-python-dotenv==1.0.1
+pynvim==0.5.0
+pyparsing==2.4.7
 pytest==8.1.1
+python-dateutil==2.8.1
+python-dotenv==1.0.1
+pytz==2020.1
+PyYAML==6.0.1
+requests==2.24.0
+rsa==4.7.2
+s3transfer==0.6.1
+scipy==1.5.2
+shellescape==3.8.1
+six==1.15.0
 SQLAlchemy==2.0.29
+sympy==1.6.2
+tabulate==0.9.0
+tomli==2.0.1
 typing_extensions==4.11.0
+urllib3==1.25.10
 Werkzeug==3.0.2
 WTForms==3.1.2
+zipp==3.18.1
