--- static/post.json
@@ -1,28 +1,0 @@
-{
-	"posts": [
-		{
-			"review": {
-				"reviewTitle": "Chu Bakery",
-				"content": "Really good croissant",
-				"image":  "https://media-cdn.tripadvisor.com/media/photo-s/0a/a9/71/fa/almond-croissant.jpg"
-				"date": "Wed Mar 25 2015 08:00:00 GMT+0800 (Australian Western Standard Time)Wed Mar 25 2015 08:00:00 GMT+0800 (Australian Western Standard Time)"
-			}
-		},
-		{
-			"review": {
-				"reviewTitle": "Chu Bakery",
-				"content": "Really good croissant",
-				"image":  "https://media-cdn.tripadvisor.com/media/photo-s/0a/a9/71/fa/almond-croissant.jpg"
-				"date": "Wed Mar 25 2015 08:00:00 GMT+0800 (Australian Western Standard Time)Wed Mar 25 2015 08:00:00 GMT+0800 (Australian Western Standard Time)"
-			}
-		},
-		{
-			"review": {
-				"reviewTitle": "Chu Bakery",
-				"content": "Really good croissant",
-				"image":  "https://media-cdn.tripadvisor.com/media/photo-s/0a/a9/71/fa/almond-croissant.jpg"
-				"date": "Wed Mar 25 2015 08:00:00 GMT+0800 (Australian Western Standard Time)Wed Mar 25 2015 08:00:00 GMT+0800 (Australian Western Standard Time)"
-			}
-		},
-	]
-}
