--- styles.css
@@ -1,84 +1,0 @@
-
-    body {
-        font-family: Arial, sans-serif;
-        margin: 0;
-        padding: 0;
-    }
-    .container {
-        display: flex;
-        height: 100vh;
-    }
-    .welcome {
-        flex: 1;
-        background-color: #f0f0f0;
-        padding: 20px;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        flex-direction: column;
-    }
-    .signup {
-        flex: 1;
-        padding: 20px;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        flex-direction: column;
-        /*height: 100%; */
-    }
-
-    .login-header-container {
-    display: none;
-    justify-content: center; /* Center horizontally */
-    align-items: center; /* Center vertically */
-}
-    
-    .login-container {
-        margin-top: 10px;
-        display: flex;
-    }
-
-    .forgotPassword-container {
-        padding: 20px;
-        justify-content: left;
-        flex-direction: column;
-    }
-
-    .login-container a {
-        display: inline-block;
-        background-color: transparent;
-        border: none;
-        color: blue;
-        text-decoration: underline; /*underline effect*/
-        cursor: pointer;
-    }
-
-    .login-container a:hover {
-        color: darkblue; /* hover colour*/
-    }
-
-    h1, p {
-        margin: 0 0 20px;
-    }
-    form {
-        background-color: #f0f0f0;
-        padding: 20px;
-        border-radius: 8px;
-        max-width: 400px;
-    }
-    input[type="text"], input[type="password"], input[type="submit"] {
-        width: 100%;
-        padding: 10px;
-        margin-bottom: 10px;
-        border: 1px solid #ccc;
-        border-radius: 4px;
-        box-sizing: border-box;
-    }
-    input[type="submit"] {
-        background-color: #4caf50;
-        color: white;
-        cursor: pointer;
-    }
-    input[type="submit"]:hover {
-        background-color: #45a049;
-    }
